requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-SQLAlchemy==1.0
SQLAlchemy==0.9.7
WebTest>=2.0.18
WTForms==2.0.1
coverage==3.7.1
pytest==2.6.3
pytest-cov==1.8.0
pytest-flask==0.4.0
python-dateutil>=2.4.2
six=>1.9.0

git+https://github.com/marshmallow-code/smore@dev
git+https://github.com/marshmallow-code/marshmallow-sqlalchemy@dev
git+https://github.com/jmcarp/marshmallow-pagination@master
